loader/requirements.txt
einops
qdrant-client
pandas
streamlit
requests
python-dotenv
